Refactor PlaylistScroller
