+++ components/playlistScroller/control.module.scss
@@ -1,0 +1,46 @@
+@import '../../variables.scss';
+
+.control {
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  width: $sideMargin;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-size: 2em;
+  color: #fff;
+  background-color: rgba(0, 0, 0, 0.3);
+  font-weight: bold;
+  opacity: 0;
+  cursor: pointer;
+  transition: all 0.2s ease;
+
+  &.prev {
+    left: 0;
+  }
+
+  &.next {
+    right: 0;
+  }
+
+  &.enabled {
+    display: flex;
+  }
+
+  &.disabled {
+    display: none;
+  }
+
+  &:hover {
+    background-color: rgba(0, 0, 0, 0.5);
+  }
+
+  &:hover .chevron {
+    transform: scale(1.2);
+  }
+
+  .chevron {
+    transition: all 0.2s ease;
+  }
+}
