+++ components/playlistScroller/index.ts
@@ -1,0 +1,3 @@
+import PlaylistScroller from './playlistScroller'
+
+export default PlaylistScroller
