+++ components/playlistScroller/item.module.scss
@@ -1,0 +1,27 @@
+@import '../../variables.scss';
+
+.item {
+  position: relative;
+  list-style: none;
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+  flex-grow: 0;
+  flex-shrink: 0;
+  background-color: $placeholderBackgroundColor;
+  margin-right: 10px;
+  cursor: pointer;
+
+  .video {
+    padding-top: 47.36%;
+    width: 100%;
+    background-position: center;
+    background-repeat: no-repeat;
+    background-size: cover;
+  }
+
+  .details {
+    .title {
+    }
+  }
+}
